+++ CHANGELOG.md
@@ -1,12 +1,17 @@
+## [0.0.3] - 2021.01.15
+* rename `authToken` to `accesToken`
+* **BREAKING:** add `fields` to customize the feed
+* refactor files
+* catch error in data body
 ## [0.0.2+1] - 2021.01.14
 * fix demo is not displayed
 
 ## [0.0.2] - 2021.01.14
 * add example
 * add Docu
-* add custom limit
-* add custom waiting widget
-* add custom noDataOrError widget
+* add custom `limit`
+* add custom `waiting` widget
+* add custom `noDataOrError` widget
 
 ## [0.0.1] - 2021.01.14
 * inital Release